requirements.txt
asgiref==3.10.0
click==8.3.0
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
mysqlclient==2.2.7
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
